Travel-Itinerary-Generator/config.py
import os
from dotenv import load_dotenv

base_dir = os.path.dirname(os.path.abspath(__file__))
env_path = os.path.join(base_dir, ".env")
load_dotenv(env_path)

MODEL_NAME = os.getenv("MODEL_NAME", "llama3.2")
